requirements.txt
adafruit-circuitpython-dotstar==2.2.9
Flask
Pillow
numpy
matplotlib
